+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2023-06-29
+
 ### Added
 
 - Revise `CloseButton` to have a more consistent look with the rest of the components.
@@ -189,7 +191,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 | `xl`  | 60px   | 48px  |
 | `2xl` | n/A    | 60px  |
 
-[unreleased]: https://github.com/chakra-ui/pro-theme/compare/0.6.0...HEAD
+[unreleased]: https://github.com/chakra-ui/pro-theme/compare/0.7.0...HEAD
 [0.1.0]: https://github.com/chakra-ui/pro-theme/releases/tag/0.1.0
 [0.1.1]: https://github.com/chakra-ui/pro-theme/releases/tag/0.1.1
 [0.2.0]: https://github.com/chakra-ui/pro-theme/releases/tag/0.2.0
@@ -199,3 +201,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.4.0]: https://github.com/chakra-ui/pro-theme/releases/tag/0.4.0
 [0.5.0]: https://github.com/chakra-ui/pro-theme/releases/tag/0.5.0
 [0.6.0]: https://github.com/chakra-ui/pro-theme/releases/tag/0.6.0
+
+[0.7.0]: https://github.com/chakra-ui/pro-theme/releases/tag/0.7.0